Expand, notify cleanup + tests (#336)
* wip

* wip: goapp builds

* more wip

* wip

* fixed examples README port ref

* removed a couple accidental changes

* wip - switching to a different task

* use functional options for better readability of the app

* a few fixes

* notify command

* tests pass

* style + comments + minor

* Eliminated Context.ActingUserID altogether

* wip stash

* wip /info works again, with IDs only

* wip example-expand user action seems to work (happy)

* Fixed expand level none parsing

* moved/rewrote all REST API tests

* goapp, and use it in REST API tests

* Makefile adjustment

* space

* moved code around

* Update apps/goapp/bindable.go


* PR feedback

* PR feedback: comment

* PR feedback: remove ServiceNow reference

* PR feedback: strings

* PR feedback: experimental

* fixed a test

* remamed invoke_call.go back to call.go to see the diff better

* make all works, moved a bunch of test stubs to restapitest

* ci to use go 1.18.2

* up golangci-lint to 1.46.1

* Updated OAuth2 REST tests

* add test for AppsMetadata returned by proxy

* bindings test

* passes rest test again

* more generic bindings test

* more bindings-related cleanup

* wip

* more bindings

* finished bindings test

* reenabled other tests

* beginning of notify test

* more refactoring

* touches

* more touches

* notify tests

* Update server/appservices/subscriptions.go


* added test for SubjectUserLeftTeam

* notify all subjects, happy, no expand

* lint

* removed scratchpad lines

* PR feedback: typo

* Update server/proxy/invoke_oauth2.go


* Update server/proxy/service.go


* test user_created with expand and all client combinations

* attempted to add test-rest-api to coverage

* Revert "attempted to add test-rest-api to coverage"

This reverts commit fcbcb4e3dbee374b6475692a0396f622d073896e.

* bot_joined_channel

* lint

* comments

* bot_left_channel + refactoring

* bot_joined/left_team

* lint

* go 1.18.3

* more ci go version

* updated mmserver from 6.6 to current

* more tests

* fixed E2E docker image back to 6.6

* fixed E2E docker image back to 6.6 once more

* lowered required server to 6.6 fot the E2E tests to pass

* restructuring tests/precision

* added bot_left_channel, more restructuring/fixes

* style

* saved a second off the test by using cached values for comparisons, where possible

* All tests pass

* type in cicrleci config

* Update server/httpin/service.go


* Update server/proxy/invoke_bindings.go


* Update server/proxy/invoke_bindings.go


* style

* fixed static and added a test

* comment

* Uninstall cleans out subscriptions

* removed old notfy tests

* added uninstall test

* Finished uninstall test

* lint

* i18n

* PR feedback and fixes

* PR feedback + TestRESTAPI/echo/email_expansion_is_controlled_by_server_settings

* renamed call.go -> invoke_call.go for consistency

* PR feedback: added comments

* fixed tests

* Examples: lifecycle, webhooks

* Example: OAuth2

* serverless example + small fixes for static

* fixed bindings test

* wip

* fixed the go.mod mess

* PR feedback: comments and strins

* set content-type json

* fixed the test app (#353)

* Update mattermost-server to 7.2, improve circle-ci caching (#354)

* fixed the test app

* Updated to server v7.2

* go mod tidy

* try release 7.1

* try 1

* try 2

* try 3

* try 4

* try 5

* try 6

* fixed embedded binding in test app (appID)

* test app update

+++ server/httpout/service.go
@@ -17,7 +17,7 @@ type Service interface {
 	config.Configurable
 	httpservice.HTTPService
 
-	GetFromURL(url string, trusted bool, limit int64) ([]byte, error)
+	GetFromURL(url string, trusted bool, limit int) ([]byte, error)
 }
 
 type service struct {
@@ -41,7 +41,7 @@ func (s *service) Configure(_ config.Config, _ utils.Logger) error {
 	return nil
 }
 
-func (s *service) GetFromURL(url string, trusted bool, limit int64) ([]byte, error) {
+func (s *service) GetFromURL(url string, trusted bool, limit int) ([]byte, error) {
 	client := s.MakeClient(trusted)
 	resp, err := client.Get(url)
 	if err != nil {
